compost: drop v1 frequency profile — single dynamic multi-view service
v1 becomes the past. The new frequency_profile_service.py IS v2:
multi-view, PPR-backed, IDF-weighted, inverse-variance fused. No
version branches, no hash prefix, no ?version= query param, no
parallel v2 module. One service, one shape.

Callers updated to the views-based shape:
- api/app/routers/graph.py — single profile/resonance/verify/sign/resonant path
- api/app/routers/concepts.py — _format_profile returns views + top
- api/app/middleware/read_tracking.py — magnitude() takes views
- api/app/services/graph_service.py — single invalidate hook
- api/tests/test_source_artifact_sensing_graph.py — asserts new shape only
- web/app/profile/[contributorId]/page.tsx — view-summary grid always on,
  explanation text is the definition, no legacy fallback

630 tests pass. tsc clean.
