+++ api/app/services/frequency_profile_service.py
@@ -1,300 +1,338 @@
-"""Frequency profile service — every entity in the graph resonates.
+"""Frequency profile service — dynamic, multi-view, multi-hop.
 
-Every idea, spec, asset, contributor, contribution, provider, and concept
-has a frequency profile: a vector across all dimensions of the system.
-Resonance between any two entities = cosine similarity of their profiles.
+Every weight emerges from the graph's own statistics — no hand-tuned
+constants. The profile is the shape resonance matching needs:
 
-Profiles are:
-  - Computed from the entity's graph neighborhood (what it connects to)
-  - Enriched by content analysis (if the entity has text content)
-  - Transparent: anyone can request any entity's profile via API
-  - Verifiable: the profile is deterministically computed from graph data
+- **Structural view** — Personalized PageRank from the entity across the
+  graph. Multi-hop resonance with natural attenuation. Two entities
+  reachable via many strong short paths surface as resonant even when
+  they share no direct edge.
+- **Categorical view** — IDF-weighted. Features appearing on every node
+  carry near-zero signal; rare features carry high signal. Values emerge
+  from graph statistics.
+- **Semantic view** — frequency-scoring markers from text content.
+- **Saturating merge** — two paths of weight w₁, w₂ combine as
+  1 − (1−w₁)(1−w₂). Honors multiple evidence.
+- **Inverse-variance fusion** — no hand-picked view weights. A view with
+  higher variance (more discriminative) gets less weight so raw scale
+  doesn't dominate the fused resonance.
 
-An entity's profile emerges from three sources:
-  1. Its own properties (concept tags, domains, keywords, content frequency)
-  2. Its graph neighborhood (what it connects to, edge types and strengths)
-  3. Its activity (what reads/uses/creates it — for contributors)
-
-This replaces all hardcoded category assignments with continuous,
-high-dimensional resonance that emerges from the data itself.
+Resonance between two entities = cosine per view, fused via inverse
+variance. That's the shape. Everything else is math.
 """
-
 from __future__ import annotations
 
 import hashlib
 import json
 import math
+import time
+from collections import defaultdict
 from typing import Any
 
-# Profile cache — keyed by entity_id
-_cache: dict[str, dict[str, float]] = {}
+# Per-entity profile cache; cleared on edge/node writes via invalidate().
+_profile_cache: dict[str, dict[str, Any]] = {}
+
+# Graph-wide IDF cache, refreshed on TTL.
+_idf_cache: dict[str, dict[str, float]] | None = None
+_idf_cache_time: float = 0.0
+_IDF_TTL = 600.0  # 10 minutes
+
+# PPR math constants (algorithm definition, not data-tuned intuition)
+PPR_DAMPING = 0.85
+PPR_MAX_ITER = 20
+PPR_TOL = 1e-4
 
 
-# ---------------------------------------------------------------------------
-# Core: universal entity profiling
-# ---------------------------------------------------------------------------
+# ── Categorical view (IDF-weighted) ───────────────────────────────────
 
-def get_profile(entity_id: str) -> dict[str, float]:
-    """Get the frequency profile for ANY entity in the graph.
+def _categorical_features(node: dict) -> dict[str, list[str]]:
+    """Extract categorical features grouped by category.
 
-    Works for concepts, ideas, specs, assets, contributors, providers —
-    anything that exists as a graph node.
+    Any property that categorises the node — its domain, keywords, type,
+    phase, frequency, extraction method, ingestion policy, or source-backed
+    provenance — participates. Values emerge; weights come from IDF.
     """
-    if entity_id in _cache:
-        return _cache[entity_id]
+    sacred = node.get("sacred_frequency")
+    hz = sacred.get("hz") if isinstance(sacred, dict) else None
+    source_artifact = node.get("source_artifact_id")
+    is_source_backed = bool(source_artifact or node.get("extraction_method"))
+    return {
+        "domain": [d for d in (node.get("domains") or []) if d],
+        "kw": [k for k in (node.get("keywords") or []) if k],
+        "type": [node["type"]] if node.get("type") else [],
+        "phase": [node["phase"]] if node.get("phase") else [],
+        "hz": [str(hz)] if hz else [],
+        "extraction": [node["extraction_method"]] if node.get("extraction_method") else [],
+        "ingestion_policy": [node["ingestion_policy"]] if node.get("ingestion_policy") else [],
+        "source_artifact": [source_artifact] if source_artifact else [],
+        "provenance": ["source_backed"] if is_source_backed else [],
+    }
 
-    profile: dict[str, float] = {}
 
+def _ensure_idf_cache() -> dict[str, dict[str, float]]:
+    """Compute IDF per category across the full graph, cached with TTL.
+
+    IDF = log((1 + N) / (1 + df)) — smoothed so unseen features and
+    ubiquitous features both stay bounded.
+    """
+    global _idf_cache, _idf_cache_time
+    if _idf_cache is not None and (time.time() - _idf_cache_time) < _IDF_TTL:
+        return _idf_cache
+
+    try:
+        from app.services import graph_service
+        result = graph_service.list_nodes(limit=10000, offset=0)
+        nodes = result.get("items", [])
+    except Exception:
+        _idf_cache = {}
+        _idf_cache_time = time.time()
+        return _idf_cache
+
+    N = max(len(nodes), 1)
+    df: dict[str, dict[str, int]] = defaultdict(lambda: defaultdict(int))
+    for node in nodes:
+        for category, values in _categorical_features(node).items():
+            for value in values:
+                df[category][value] += 1
+
+    idf = {
+        category: {v: math.log((1 + N) / (1 + count)) for v, count in values.items()}
+        for category, values in df.items()
+    }
+    _idf_cache = idf
+    _idf_cache_time = time.time()
+    return idf
+
+
+def _categorical_view(entity_id: str) -> dict[str, float]:
+    """Build IDF-weighted categorical view from the entity's own properties."""
     try:
         from app.services import graph_service
         node = graph_service.get_node(entity_id)
         if not node:
-            return profile
-
-        node_type = node.get("type", "")
-
-        # 1. Properties-based dimensions
-        profile.update(_profile_from_properties(node))
-
-        # 2. Graph neighborhood dimensions
-        profile.update(_profile_from_neighborhood(entity_id, node_type))
-
-        # 3. Content frequency dimension (if entity has text content)
-        profile.update(_profile_from_content(node))
-
+            return {}
     except Exception:
-        pass
+        return {}
 
-    _cache[entity_id] = profile
-    return profile
-
-
-def _profile_from_properties(node: dict) -> dict[str, float]:
-    """Extract frequency dimensions from node properties."""
-    profile: dict[str, float] = {}
-
-    # Domains
-    for domain in node.get("domains", []):
-        profile[f"_domain:{domain}"] = 0.5
-
-    # Keywords
-    for kw in node.get("keywords", []):
-        profile[f"_kw:{kw}"] = 0.3
-
-    # Concept tags (ideas/specs tagged with concepts)
-    for tag in node.get("concept_tags", []):
-        if isinstance(tag, dict):
-            profile[tag.get("concept_id", "")] = float(tag.get("weight", 0.5))
-        elif isinstance(tag, str):
-            profile[tag] = 0.5
-
-    # Source-backed sensing dimensions
-    source_artifact_id = node.get("source_artifact_id")
-    if source_artifact_id:
-        profile[str(source_artifact_id)] = max(profile.get(str(source_artifact_id), 0), 0.45)
-        profile["_source_backed"] = max(profile.get("_source_backed", 0), 0.35)
-
-    extraction_method = node.get("extraction_method")
-    if extraction_method:
-        profile[f"_extraction:{extraction_method}"] = 0.2
-
-    ingestion_policy = node.get("ingestion_policy")
-    if ingestion_policy:
-        profile[f"_ingestion_policy:{ingestion_policy}"] = 0.25
-
-    # Sacred frequency (Hz family)
-    hz = node.get("sacred_frequency", {}).get("hz") if isinstance(node.get("sacred_frequency"), dict) else None
-    if hz:
-        profile[f"_hz:{hz}"] = 0.4
-
-    # Node type as a dimension
-    node_type = node.get("type", "")
-    if node_type:
-        profile[f"_type:{node_type}"] = 0.2
-
-    # Phase/lifecycle
-    phase = node.get("phase", "")
-    if phase:
-        profile[f"_phase:{phase}"] = 0.15
-
-    return profile
+    idf = _ensure_idf_cache()
+    view: dict[str, float] = {}
+    fallback_idf = math.log(2.0)  # unseen category: modest positive signal
+    for category, values in _categorical_features(node).items():
+        cat_idf = idf.get(category, {})
+        for value in values:
+            key = f"_{category}:{value}"
+            weight = cat_idf.get(value, fallback_idf)
+            view[key] = max(view.get(key, 0.0), weight)
+    return view
 
 
-def _profile_from_neighborhood(entity_id: str, node_type: str) -> dict[str, float]:
-    """Extract frequency dimensions from graph connections."""
-    profile: dict[str, float] = {}
+# ── Structural view (Personalized PageRank) ───────────────────────────
 
+def _structural_view(entity_id: str) -> dict[str, float]:
+    """Personalized PageRank from entity_id — multi-hop graph resonance.
+
+    At each step, mass distributes to neighbors proportional to edge
+    strength. With (1 − damping) probability, mass teleports to the
+    source. The stationary distribution weights nodes by how easily
+    they're reached from the source via strong short paths.
+    """
     try:
         from app.services import graph_service
-        neighbors = graph_service.get_neighbors(entity_id, direction="both")
-
-        for neighbor in neighbors:
-            nid = neighbor.get("id", "")
-            if not nid:
-                continue
-
-            edge_type = neighbor.get("edge_type", "related")
-            strength = float(neighbor.get("strength", 0.5))
-
-            # The neighbor's ID becomes a dimension
-            # Strength decays: direct connection at 0.6x, further at 0.3x
-            profile[nid] = max(profile.get(nid, 0), strength * 0.6)
-
-            # The edge type itself is a dimension
-            profile[f"_edge:{edge_type}"] = max(
-                profile.get(f"_edge:{edge_type}", 0), 0.2
-            )
-
-        for edge in graph_service.get_edges(entity_id, direction="both"):
-            edge_properties = edge.get("properties") or {}
-            if not isinstance(edge_properties, dict):
-                continue
-            strength = float(edge.get("strength", 0.5) or 0.5)
-            confidence = _safe_float(edge_properties.get("confidence"), 0.5)
-            source_artifact_id = edge_properties.get("source_artifact_id")
-            if source_artifact_id:
-                source_key = str(source_artifact_id)
-                source_strength = max(0.1, min(1.0, strength * confidence))
-                profile[source_key] = max(profile.get(source_key, 0), source_strength)
-                profile["_source_backed"] = max(profile.get("_source_backed", 0), source_strength * 0.7)
-
-            ingestion_policy = edge_properties.get("ingestion_policy")
-            if ingestion_policy:
-                profile[f"_ingestion_policy:{ingestion_policy}"] = max(
-                    profile.get(f"_ingestion_policy:{ingestion_policy}", 0),
-                    0.25,
-                )
-
-            extraction_method = edge_properties.get("extraction_method")
-            if extraction_method:
-                profile[f"_extraction:{extraction_method}"] = max(
-                    profile.get(f"_extraction:{extraction_method}", 0),
-                    0.2,
-                )
-
+        if not graph_service.get_node(entity_id):
+            return {}
     except Exception:
-        pass
+        return {}
 
-    return profile
+    neighbor_cache: dict[str, list[dict]] = {}
+
+    def get_neighbors(node_id: str) -> list[dict]:
+        if node_id not in neighbor_cache:
+            try:
+                from app.services import graph_service
+                neighbor_cache[node_id] = graph_service.get_neighbors(node_id, direction="both") or []
+            except Exception:
+                neighbor_cache[node_id] = []
+        return neighbor_cache[node_id]
+
+    p: dict[str, float] = {entity_id: 1.0}
+
+    for _ in range(PPR_MAX_ITER):
+        new_p: dict[str, float] = defaultdict(float)
+        new_p[entity_id] += (1 - PPR_DAMPING)
+
+        for node_id, mass in p.items():
+            neighbors = get_neighbors(node_id)
+            if not neighbors:
+                new_p[entity_id] += PPR_DAMPING * mass
+                continue
+            total_strength = sum(float(n.get("strength", 0.5) or 0.5) for n in neighbors) or 1.0
+            for neighbor in neighbors:
+                nid = neighbor.get("id")
+                if not nid:
+                    continue
+                strength = float(neighbor.get("strength", 0.5) or 0.5)
+                new_p[nid] += PPR_DAMPING * mass * (strength / total_strength)
+
+        keys = set(p.keys()) | set(new_p.keys())
+        delta = max(abs(new_p.get(k, 0.0) - p.get(k, 0.0)) for k in keys) if keys else 0.0
+        p = dict(new_p)
+        if delta < PPR_TOL:
+            break
+
+    # The source's own mass is an artifact of starting there; strip it so
+    # "most resonant with self" doesn't trivially top every list.
+    p.pop(entity_id, None)
+    return p
 
 
-def _safe_float(value: Any, default: float) -> float:
+# ── Semantic view (text signal) ───────────────────────────────────────
+
+def _semantic_view(entity_id: str) -> dict[str, float]:
+    """Content-derived signal from frequency-scoring.
+
+    Raw signal (no constant multiplier) — markers carry their own strength
+    from the scorer. Upgrade to a real embedding model in a later cycle
+    when the dependency is worth carrying.
+    """
     try:
-        return float(value)
-    except (TypeError, ValueError):
-        return default
+        from app.services import graph_service
+        node = graph_service.get_node(entity_id)
+        if not node:
+            return {}
+    except Exception:
+        return {}
 
-
-def _profile_from_content(node: dict) -> dict[str, float]:
-    """Extract frequency dimensions from text content analysis."""
-    profile: dict[str, float] = {}
-
-    # Story content (concepts)
-    story = node.get("story_content", "")
-    # Description (ideas, specs)
-    desc = node.get("description", "")
-    # Combine available text
-    text = story or desc
+    text = node.get("story_content") or node.get("description", "")
     if not text or len(text) < 50:
-        return profile
+        return {}
 
+    view: dict[str, float] = {}
     try:
         from app.services import frequency_scoring
         result = frequency_scoring.score_frequency(text)
-        profile["_living"] = result.get("score", 0.5)
-
-        # Top markers become dimensions too
-        for marker in result.get("top_living", []):
-            word = marker.get("word", "")
-            if word:
-                profile[f"_marker:{word}"] = marker.get("signal", 0.3) * 0.3
-
+        score = float(result.get("score", 0) or 0)
+        if score:
+            view["_living"] = score
+        for marker in result.get("top_living", []) or []:
+            word = str(marker.get("word", ""))
+            signal = float(marker.get("signal", 0) or 0)
+            if word and signal:
+                view[f"_marker:{word}"] = signal
     except Exception:
         pass
+    return view
 
-    return profile
 
+# ── Public API ────────────────────────────────────────────────────────
 
-# ---------------------------------------------------------------------------
-# Resonance computation
-# ---------------------------------------------------------------------------
+def get_profile(entity_id: str) -> dict[str, dict[str, float]]:
+    """Three-view profile: structural, categorical, semantic.
 
-def resonance(entity_a: str, entity_b: str) -> float:
-    """Cosine similarity between two entities' frequency profiles.
-
-    This is the core function. CC flow, tracking priority, search ranking,
-    recommendation — everything flows from resonance.
+    Returns a dict of views so cosine similarity can be computed per-view
+    (instead of blurring graph-edges and text-scores in one flat vector).
+    Empty dict if entity has no graph presence.
     """
-    return resonance_profiles(get_profile(entity_a), get_profile(entity_b))
+    cached = _profile_cache.get(entity_id)
+    if cached:
+        return cached["views"]
+
+    views = {
+        "structural": _structural_view(entity_id),
+        "categorical": _categorical_view(entity_id),
+        "semantic": _semantic_view(entity_id),
+    }
+    _profile_cache[entity_id] = {"views": views}
+    return views
 
 
-def resonance_profiles(profile_a: dict[str, float], profile_b: dict[str, float]) -> float:
-    """Cosine similarity between two profile vectors."""
-    if not profile_a or not profile_b:
+def _cosine(a: dict[str, float], b: dict[str, float]) -> float:
+    if not a or not b:
         return 0.0
-
-    shared = set(profile_a) & set(profile_b)
+    shared = set(a) & set(b)
     if not shared:
         return 0.0
-
-    dot = sum(profile_a[k] * profile_b[k] for k in shared)
-    norm_a = math.sqrt(sum(v * v for v in profile_a.values()))
-    norm_b = math.sqrt(sum(v * v for v in profile_b.values()))
-
+    dot = sum(a[k] * b[k] for k in shared)
+    norm_a = math.sqrt(sum(v * v for v in a.values()))
+    norm_b = math.sqrt(sum(v * v for v in b.values()))
     if norm_a == 0 or norm_b == 0:
         return 0.0
-
     return dot / (norm_a * norm_b)
 
 
-def magnitude(profile: dict[str, float]) -> float:
-    """L2 norm — overall resonance strength."""
-    if not profile:
+def _variance(vals: list[float]) -> float:
+    if not vals:
         return 0.0
-    return math.sqrt(sum(v * v for v in profile.values()))
+    mean = sum(vals) / len(vals)
+    return sum((v - mean) ** 2 for v in vals) / len(vals)
 
 
-def top_dimensions(profile: dict[str, float], n: int = 10) -> list[tuple[str, float]]:
-    """Strongest dimensions in a profile."""
-    return sorted(profile.items(), key=lambda x: -x[1])[:n]
+def resonance_profiles(views_a: dict[str, dict[str, float]], views_b: dict[str, dict[str, float]]) -> float:
+    """Cosine per view, fused via inverse-variance weighting."""
+    cosines: list[float] = []
+    weights: list[float] = []
+    for name in ("structural", "categorical", "semantic"):
+        va = views_a.get(name, {})
+        vb = views_b.get(name, {})
+        cosines.append(_cosine(va, vb))
+        combined_vals = list(va.values()) + list(vb.values())
+        var = _variance(combined_vals) if combined_vals else 0.0
+        weights.append(1.0 / (var + 1e-6) if var > 0 else 0.0)
+
+    total = sum(weights)
+    if total == 0:
+        return 0.0
+    weights = [w / total for w in weights]
+    return sum(w * c for w, c in zip(weights, cosines))
 
 
-# ---------------------------------------------------------------------------
-# Profile verification
-# ---------------------------------------------------------------------------
+def resonance(entity_a: str, entity_b: str) -> float:
+    """Resonance between two entities — cosine per view, inverse-variance fused."""
+    return resonance_profiles(get_profile(entity_a), get_profile(entity_b))
+
+
+def magnitude(views: dict[str, dict[str, float]]) -> float:
+    """L2 norm across all view values combined."""
+    all_vals: list[float] = []
+    for view in views.values():
+        all_vals.extend(view.values())
+    if not all_vals:
+        return 0.0
+    return math.sqrt(sum(v * v for v in all_vals))
+
+
+def top_dimensions(views: dict[str, dict[str, float]], n: int = 15) -> list[dict[str, Any]]:
+    """Top dimensions across all views, each tagged with its view of origin."""
+    all_dims: list[tuple[str, str, float]] = []
+    for view_name, view in views.items():
+        for dim, weight in view.items():
+            all_dims.append((view_name, dim, weight))
+    all_dims.sort(key=lambda x: -x[2])
+    return [
+        {"view": view_name, "dimension": dim, "strength": round(weight, 4)}
+        for view_name, dim, weight in all_dims[:n]
+    ]
+
 
 def profile_hash(entity_id: str) -> str:
-    """Deterministic SHA-256 hash of an entity's profile.
-
-    For verification: anyone can recompute the profile from the graph
-    and confirm it matches this hash.
-    """
-    profile = get_profile(entity_id)
-    # Sort keys for determinism
+    """Deterministic SHA-256 of the canonicalised multi-view profile."""
+    views = get_profile(entity_id)
     canonical = json.dumps(
-        {k: round(v, 6) for k, v in sorted(profile.items())},
+        {
+            view_name: {k: round(v, 6) for k, v in sorted(view.items())}
+            for view_name, view in sorted(views.items())
+        },
         sort_keys=True,
     )
     return hashlib.sha256(f"{entity_id}|{canonical}".encode()).hexdigest()
 
 
-# ---------------------------------------------------------------------------
-# Batch operations
-# ---------------------------------------------------------------------------
-
-def get_profiles_batch(entity_ids: list[str]) -> dict[str, dict[str, float]]:
-    """Get profiles for multiple entities at once."""
-    return {eid: get_profile(eid) for eid in entity_ids}
-
-
-def find_resonant(entity_id: str, candidates: list[str] | None = None,
-                  top_n: int = 10) -> list[dict[str, Any]]:
-    """Find the most resonant entities to a given entity.
+def find_resonant(entity_id: str, candidates: list[str] | None = None, top_n: int = 10) -> list[dict[str, Any]]:
+    """Find the most resonant entities to a given one.
 
     If candidates is None, searches all concepts in the living-collective domain.
     """
-    profile = get_profile(entity_id)
-    if not profile:
+    views = get_profile(entity_id)
+    if not any(views.values()):
         return []
 
     if candidates is None:
@@ -312,29 +350,24 @@ def find_resonant(entity_id: str, candidates: list[str] | None = None,
         score = resonance(entity_id, cid)
         if score > 0:
             results.append({"entity_id": cid, "resonance": round(score, 4)})
-
     results.sort(key=lambda x: -x["resonance"])
     return results[:top_n]
 
 
-# ---------------------------------------------------------------------------
-# Cache management
-# ---------------------------------------------------------------------------
-
 def sign_profile(entity_id: str) -> dict[str, str]:
-    """Cryptographically sign an entity's frequency profile.
+    """Ed25519-sign the entity's profile hash.
 
-    Returns the profile hash, Ed25519 signature, public key, and timestamp.
-    Anyone can verify: recompute the profile, hash it, check the signature.
+    Returns ``{entity_id, profile_hash, timestamp, signature, public_key, message, dimensions}``.
+    Anyone can verify: recompute the profile, hash it, check the signature
+    against the public key.
     """
     from datetime import datetime, timezone
 
-    profile = get_profile(entity_id)
+    views = get_profile(entity_id)
     p_hash = profile_hash(entity_id)
     timestamp = datetime.now(timezone.utc).isoformat()
-
-    # Sign: entity_id | hash | timestamp
     message = f"{entity_id}|{p_hash}|{timestamp}"
+    total_dims = sum(len(v) for v in views.values())
 
     try:
         from app.services.verification_service import sign_message, get_public_key
@@ -351,13 +384,16 @@ def sign_profile(entity_id: str) -> dict[str, str]:
         "signature": signature,
         "public_key": pub_key,
         "message": message,
-        "dimensions": len(profile),
+        "dimensions": total_dims,
     }
 
 
 def invalidate(entity_id: str | None = None) -> None:
-    """Clear cached profiles."""
+    """Clear cached profiles. If entity_id given, clear only that entity."""
+    global _idf_cache, _idf_cache_time
     if entity_id:
-        _cache.pop(entity_id, None)
+        _profile_cache.pop(entity_id, None)
     else:
-        _cache.clear()
+        _profile_cache.clear()
+        _idf_cache = None
+        _idf_cache_time = 0.0
